zoza/requirements.txt
openai
pylint
pytest
rich
icecream
python-dotenv